Refactored to reduce boilerplate and make single Impl
