--- java/src/main/java/com/amido/stacks/workloads/menu/api/v1/dto/response/GenerateTokenResponse.java
@@ -1,21 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class GenerateTokenResponse {
-
-  @JsonProperty("access_token")
-  private String access_token = null;
-
-  @JsonProperty("expires_in")
-  private String expires_in = null;
-
-  @JsonProperty("token_type")
-  private String token_type = null;
-}
