--- java/src/main/java/com/amido/stacks/workloads/menu/api/v1/impl/CreateItemControllerImpl.java
@@ -1,22 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
-
-import com.amido.stacks.workloads.menu.api.v1.CreateItemController;
-import com.amido.stacks.workloads.menu.api.v1.dto.request.CreateItemRequest;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.ResourceCreatedResponse;
-import java.util.UUID;
-import javax.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class CreateItemControllerImpl implements CreateItemController {
-
-  public CreateItemControllerImpl() {}
-
-  @Override
-  public ResponseEntity<ResourceCreatedResponse> addMenuItem(
-      UUID menuId, UUID categoryId, @Valid CreateItemRequest body, String correlationId) {
-    return new ResponseEntity<>(new ResourceCreatedResponse(UUID.randomUUID()), HttpStatus.CREATED);
-  }
-}
