--- java/src/main/java/com/amido/stacks/workloads/menu/api/v1/impl/CreateMenuControllerImpl.java
@@ -1,22 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
-
-import com.amido.stacks.workloads.menu.api.v1.CreateMenuController;
-import com.amido.stacks.workloads.menu.api.v1.dto.request.CreateMenuRequest;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.ResourceCreatedResponse;
-import java.util.UUID;
-import javax.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class CreateMenuControllerImpl implements CreateMenuController {
-
-  public CreateMenuControllerImpl() {}
-
-  @Override
-  public ResponseEntity<ResourceCreatedResponse> createMenu(
-      @Valid CreateMenuRequest body, String correlationId) {
-    return new ResponseEntity<>(new ResourceCreatedResponse(UUID.randomUUID()), HttpStatus.CREATED);
-  }
-}
