--- java/src/main/java/com/amido/stacks/workloads/menu/api/v1/impl/DeleteMenuControllerImpl.java
@@ -1,24 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
-
-import static org.springframework.http.HttpStatus.OK;
-
-import com.amido.stacks.workloads.menu.api.v1.DeleteMenuController;
-import java.util.UUID;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * DeleteMenuController implementation.
- *
- * @author ArathyKrishna
- */
-@RestController
-public class DeleteMenuControllerImpl implements DeleteMenuController {
-
-  public DeleteMenuControllerImpl() {}
-
-  @Override
-  public ResponseEntity<Void> deleteMenu(UUID menuId, String correlationId) {
-    return new ResponseEntity<>(OK);
-  }
-}
