--- java/src/main/java/com/amido/stacks/workloads/menu/api/v1/impl/QueryMenuControllerImpl.java
@@ -1,81 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
-
-import com.amido.stacks.workloads.menu.api.v1.QueryMenuController;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.MenuDTO;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.SearchMenuResult;
-import com.amido.stacks.workloads.menu.domain.Category;
-import com.amido.stacks.workloads.menu.domain.Item;
-import com.amido.stacks.workloads.menu.domain.Menu;
-import com.amido.stacks.workloads.menu.mappers.MenuMapper;
-import com.amido.stacks.workloads.menu.mappers.SearchMenuResultItemMapper;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.UUID;
-import java.util.stream.Collectors;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RestController;
-
-/** MenuControllerImpl - MenuDTO Controller used to interact and manage menus API. */
-@RestController
-public class QueryMenuControllerImpl implements QueryMenuController {
-
-  Logger logger = LoggerFactory.getLogger(QueryMenuControllerImpl.class);
-
-  @Autowired private MenuMapper menuMapper;
-
-  @Autowired private SearchMenuResultItemMapper searchMenuResultItemMapper;
-
-  @Override
-  public ResponseEntity<SearchMenuResult> searchMenu(
-      String searchTerm, UUID restaurantId, Integer pageSize, Integer pageNumber) {
-
-    List<Menu> menuList = new ArrayList<>();
-
-    final String menuId = "d5785e28-306b-4e23-add0-3f9092d395f8";
-
-    Menu mockMenu;
-    if (restaurantId == null) {
-      mockMenu =
-          new Menu(
-              menuId,
-              "58a1df85-6bdc-412a-a118-0f0e394c1342",
-              "name",
-              "description",
-              new ArrayList<>(),
-              true);
-    } else {
-      mockMenu =
-          new Menu(menuId, restaurantId.toString(), "name", "description", new ArrayList<>(), true);
-    }
-
-    menuList.add(mockMenu);
-
-    return ResponseEntity.ok(
-        new SearchMenuResult(
-            pageSize,
-            pageNumber,
-            menuList.stream()
-                .map(m -> searchMenuResultItemMapper.toDto(m))
-                .collect(Collectors.toList())));
-  }
-
-  @Override
-  public ResponseEntity<MenuDTO> getMenu(UUID id, String correlationId) {
-    final String restaurantId = "58a1df85-6bdc-412a-a118-0f0e394c1342";
-    final String categoryId = "2c43dbda-7d4d-46fb-b246-bec2bd348ca1";
-    final String itemId = "7e46a698-080b-45e6-a529-2c196d00791c";
-
-    Menu menu =
-        new Menu(id.toString(), restaurantId, "name", "description", new ArrayList<>(), true);
-    Item item = new Item(itemId, "item name", "item description", 5.99d, true);
-    Category category =
-        new Category(categoryId, "cat name", "cat description", Arrays.asList(item));
-    menu.addOrUpdateCategory(category);
-
-    return ResponseEntity.ok(menuMapper.toDto(menu));
-  }
-}
