+++ java/src/main/java/com/amido/stacks/workloads/menu/api/v2/MenuControllerV2.java
@@ -2,25 +2,29 @@ package com.amido.stacks.workloads.menu.api.v2;
 
 import com.amido.stacks.core.api.annotations.ReadAPIResponses;
 import com.amido.stacks.workloads.menu.api.v1.dto.response.MenuDTO;
+import com.amido.stacks.workloads.menu.domain.Menu;
+import com.amido.stacks.workloads.menu.mappers.MenuMapper;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import java.util.ArrayList;
 import java.util.UUID;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
-@RequestMapping(
-    path = "/v2/menu",
-    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
-    method = RequestMethod.GET)
-public interface QueryMenuControllerV2 {
+@RequestMapping(path = "/v2/menu", produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8")
+@RestController
+public class MenuControllerV2 {
+
+  @Autowired private MenuMapper menuMapper;
 
   @GetMapping(value = "/{id}")
   @Operation(
@@ -39,5 +43,14 @@ public interface QueryMenuControllerV2 {
   @ReadAPIResponses
   ResponseEntity<MenuDTO> getMenu(
       @PathVariable(name = "id") UUID id,
-      @Parameter(hidden = true) @RequestAttribute("CorrelationId") String correlationId);
+      @Parameter(hidden = true) @RequestAttribute("CorrelationId") String correlationId) {
+
+    String restaurantId = "3930ddff-82ce-4f7e-b910-b0709b276cf0";
+
+    Menu menu =
+        new Menu(
+            id.toString(), restaurantId, "0 Menu", "0 Menu Description", new ArrayList<>(), true);
+
+    return ResponseEntity.ok(menuMapper.toDto(menu));
+  }
 }
