+++ java/src/test/java/com/amido/stacks/workloads/menu/api/v1/CategoryControllerTest.java
@@ -1,17 +1,21 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
+package com.amido.stacks.workloads.menu.api.v1;
 
 import static com.amido.stacks.workloads.menu.domain.CategoryHelper.createCategories;
 import static com.amido.stacks.workloads.menu.domain.CategoryHelper.createCategory;
+import static com.amido.stacks.workloads.menu.domain.ItemHelper.createItem;
 import static com.amido.stacks.workloads.menu.domain.MenuHelper.createMenu;
 import static com.amido.stacks.workloads.util.TestHelper.getBaseURL;
 import static com.amido.stacks.workloads.util.TestHelper.getRequestHttpEntity;
 import static java.util.UUID.fromString;
 import static org.assertj.core.api.BDDAssertions.then;
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.OK;
 
 import com.amido.stacks.core.api.dto.ErrorResponse;
+import com.amido.stacks.core.api.dto.response.ResourceCreatedResponse;
+import com.amido.stacks.core.api.dto.response.ResourceUpdatedResponse;
+import com.amido.stacks.workloads.menu.api.v1.dto.request.CreateCategoryRequest;
 import com.amido.stacks.workloads.menu.api.v1.dto.request.UpdateCategoryRequest;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.ResourceUpdatedResponse;
 import com.amido.stacks.workloads.menu.domain.Category;
 import com.amido.stacks.workloads.menu.domain.Menu;
 import java.util.List;
@@ -24,19 +28,62 @@ import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @EnableAutoConfiguration
 @Tag("Integration")
 @ActiveProfiles("test")
-class UpdateCategoryControllerImplTest {
+public class CategoryControllerTest {
+
+  public static final String CREATE_CATEGORY = "%s/v1/menu/%s/category";
 
   public static final String UPDATE_CATEGORY = "%s/v1/menu/%s/category/%s";
+
+  public static final String DELETE_CATEGORY = "%s/v1/menu/%s/category/%s";
 
   @LocalServerPort private int port;
 
   @Autowired private TestRestTemplate testRestTemplate;
+
+  @Test
+  void testInvalidMenuIdWilThrowBadRequest() {
+    // Given
+    CreateCategoryRequest request = new CreateCategoryRequest("test Category Name", "");
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            String.format(CREATE_CATEGORY, getBaseURL(port), "XXXXXX"),
+            request,
+            ErrorResponse.class);
+
+    // Then
+    then(response).isNotNull();
+    then(response.getStatusCode()).isEqualTo(BAD_REQUEST);
+  }
+
+  @Test
+  void testAddCategory() {
+    // Given
+    Menu menu = createMenu(1);
+
+    CreateCategoryRequest request =
+        new CreateCategoryRequest("test Category Name", "test Category Description");
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            String.format(CREATE_CATEGORY, getBaseURL(port), menu.getId()),
+            request,
+            ResourceCreatedResponse.class);
+
+    // Then
+    then(response).isNotNull();
+    then(response.getBody()).isNotNull();
+    then(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
+  }
 
   @Test
   void testUpdateCategorySuccess() {
@@ -118,6 +165,70 @@ class UpdateCategoryControllerImplTest {
 
     // Then
     then(response).isNotNull();
+    then(response.getStatusCode()).isEqualTo(OK);
+  }
+
+  @Test
+  void testDeleteCategorySuccess() {
+    // Given
+    Menu menu = createMenu(1);
+    Category category = createCategory(0);
+    menu.setCategories(List.of(category));
+
+    // When
+    String requestUrl =
+        String.format(DELETE_CATEGORY, getBaseURL(port), menu.getId(), category.getId());
+    var response =
+        this.testRestTemplate.exchange(
+            requestUrl,
+            HttpMethod.DELETE,
+            new HttpEntity<>(getRequestHttpEntity()),
+            ErrorResponse.class);
+
+    // Then
+    then(response.getStatusCode()).isEqualTo(OK);
+  }
+
+  @Test
+  void testDeleteCategoryWithAnItem() {
+    // Given
+    Menu menu = createMenu(1);
+    Category category = createCategory(0);
+    category.addOrUpdateItem(createItem(0));
+    menu.addOrUpdateCategory(category);
+
+    // When
+    String requestUrl =
+        String.format(DELETE_CATEGORY, getBaseURL(port), menu.getId(), category.getId());
+    var response =
+        this.testRestTemplate.exchange(
+            requestUrl,
+            HttpMethod.DELETE,
+            new HttpEntity<>(getRequestHttpEntity()),
+            ErrorResponse.class);
+
+    // Then
+    then(response.getStatusCode()).isEqualTo(OK);
+  }
+
+  @Test
+  void testDeleteACategoryFromList() {
+    // Given
+    Menu menu = createMenu(1);
+    List<Category> categories = createCategories(2);
+    menu.setCategories(categories);
+
+    // When
+    String requestUrl =
+        String.format(DELETE_CATEGORY, getBaseURL(port), menu.getId(), categories.get(0).getId());
+    var response =
+        this.testRestTemplate.exchange(
+            requestUrl,
+            HttpMethod.DELETE,
+            new HttpEntity<>(getRequestHttpEntity()),
+            ErrorResponse.class);
+
+    // Then
     then(response.getStatusCode()).isEqualTo(OK);
   }
 }
