+++ java/src/test/java/com/amido/stacks/workloads/menu/api/v1/MenuControllerTest.java
@@ -1,12 +1,19 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
+package com.amido.stacks.workloads.menu.api.v1;
 
+import static com.amido.stacks.workloads.menu.domain.MenuHelper.createMenu;
 import static com.amido.stacks.workloads.menu.domain.MenuHelper.createMenus;
 import static com.amido.stacks.workloads.util.TestHelper.getBaseURL;
+import static com.amido.stacks.workloads.util.TestHelper.getRequestHttpEntity;
 import static org.assertj.core.api.BDDAssertions.then;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.core.IsNull.notNullValue;
+import static org.springframework.http.HttpStatus.OK;
 
+import com.amido.stacks.core.api.dto.response.ResourceCreatedResponse;
+import com.amido.stacks.core.api.dto.response.ResourceUpdatedResponse;
+import com.amido.stacks.workloads.menu.api.v1.dto.request.CreateMenuRequest;
+import com.amido.stacks.workloads.menu.api.v1.dto.request.UpdateMenuRequest;
 import com.amido.stacks.workloads.menu.api.v1.dto.response.MenuDTO;
 import com.amido.stacks.workloads.menu.api.v1.dto.response.SearchMenuResult;
 import com.amido.stacks.workloads.menu.api.v1.dto.response.SearchMenuResultItem;
@@ -21,6 +28,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.UUID;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
@@ -29,14 +37,26 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @EnableAutoConfiguration
 @Tag("Integration")
 @ActiveProfiles("test")
-public class QueryMenuControllerImplTest {
+public class MenuControllerTest {
+
+  public static final String CREATE_MENU = "/v1/menu";
+
+  public static final String UPDATE_MENU = "%s/v1/menu/%s";
+
+  public static final String DELETE_MENU = "%s/v1/menu/%s";
+
+  final int DEFAULT_PAGE_NUMBER = 1;
+  final int DEFAULT_PAGE_SIZE = 20;
 
   @LocalServerPort private int port;
 
@@ -50,8 +70,22 @@ public class QueryMenuControllerImplTest {
 
   @Autowired private SearchMenuResultItemMapper searchMenuResultItemMapper;
 
-  final int DEFAULT_PAGE_NUMBER = 1;
-  final int DEFAULT_PAGE_SIZE = 20;
+  @Test
+  void testCreateNewMenu() {
+    // Given
+    Menu m = createMenu(1);
+    CreateMenuRequest request =
+        new CreateMenuRequest(
+            m.getName(), m.getDescription(), UUID.fromString(m.getRestaurantId()), m.getEnabled());
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            getBaseURL(port) + CREATE_MENU, request, ResourceCreatedResponse.class);
+
+    // Then
+    then(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
+  }
 
   @Test
   public void listMenusAndPagination() {
@@ -146,5 +180,40 @@ public class QueryMenuControllerImplTest {
     assertThat(actual, is(notNullValue()));
     assertThat(actual.getPageNumber(), is(DEFAULT_PAGE_NUMBER));
     assertThat(actual.getPageSize(), is(DEFAULT_PAGE_SIZE));
+  }
+
+  @Test
+  void testUpdateSuccess() {
+    // Given
+    Menu menu = createMenu(0);
+
+    UpdateMenuRequest request = new UpdateMenuRequest("new name", "new description", false);
+
+    // When
+    var response =
+        this.testRestTemplate.exchange(
+            String.format(UPDATE_MENU, getBaseURL(port), menu.getId()),
+            HttpMethod.PUT,
+            new HttpEntity<>(request, getRequestHttpEntity()),
+            ResourceUpdatedResponse.class);
+
+    // Then
+    then(response).isNotNull();
+    then(response.getStatusCode()).isEqualTo(HttpStatus.OK);
+  }
+
+  @Test
+  void testDeleteMenuSuccess() {
+    // Given
+    Menu menu = createMenu(1);
+
+    var response =
+        this.testRestTemplate.exchange(
+            String.format(DELETE_MENU, getBaseURL(port), menu.getId()),
+            HttpMethod.DELETE,
+            new HttpEntity<>(getRequestHttpEntity()),
+            ResponseEntity.class);
+    // Then
+    then(response.getStatusCode()).isEqualTo(OK);
   }
 }
